+++ ios/InputPilotTests/TrackpadGesturesTests.swift
@@ -42,8 +42,49 @@ final class TrackpadGesturesTests: XCTestCase {
 
     func testZoomContributionZoomsInOnScaleUp() {
         XCTAssertEqual(TrackpadGestures.zoomContribution(scaleChange: 1.0), 0, accuracy: 0.0001)
-        XCTAssertEqual(TrackpadGestures.zoomContribution(scaleChange: 1.1), 1, accuracy: 0.0001)
-        XCTAssertEqual(TrackpadGestures.zoomContribution(scaleChange: 0.9), -1, accuracy: 0.0001)
+        XCTAssertEqual(TrackpadGestures.zoomContribution(scaleChange: 1.1), 0.1 * TrackpadGestures.zoomLinesPerFullScale, accuracy: 0.0001)
+        XCTAssertEqual(TrackpadGestures.zoomContribution(scaleChange: 0.9), -0.1 * TrackpadGestures.zoomLinesPerFullScale, accuracy: 0.0001)
+    }
+
+    func testDragEngagementRequiresSustainedMovement() {
+        XCTAssertFalse(TrackpadGestures.isDragEngagement(dx: 3, dy: 4))
+        XCTAssertFalse(TrackpadGestures.isDragEngagement(dx: 9, dy: 0))
+        XCTAssertTrue(TrackpadGestures.isDragEngagement(dx: 6, dy: 8))
+        XCTAssertTrue(TrackpadGestures.isDragEngagement(dx: 0, dy: -12))
+    }
+
+    func testTwoFingerArbiterLocksScrollOnPurePan() {
+        var arbiter = TwoFingerArbiter(centroidX: 0, centroidY: 0, distance: 100)
+        XCTAssertNil(arbiter.update(centroidX: 5, centroidY: 5, distance: 102))
+        XCTAssertNil(arbiter.update(centroidX: 11, centroidY: 0, distance: 101))
+        XCTAssertEqual(arbiter.update(centroidX: 13, centroidY: 0, distance: 100), .scroll)
+        XCTAssertEqual(arbiter.mode, .scroll)
+        // Once locked, later geometry can never switch the mode to zoom.
+        XCTAssertNil(arbiter.update(centroidX: 40, centroidY: 0, distance: 60))
+        XCTAssertEqual(arbiter.mode, .scroll)
+    }
+
+    func testTwoFingerArbiterLocksZoomOnPinch() {
+        var arbiter = TwoFingerArbiter(centroidX: 0, centroidY: 0, distance: 100)
+        XCTAssertNil(arbiter.update(centroidX: 3, centroidY: 2, distance: 104))
+        XCTAssertEqual(arbiter.update(centroidX: 4, centroidY: 2, distance: 111), .zoom)
+        XCTAssertEqual(arbiter.mode, .zoom)
+        // Once locked, later centroid movement can never switch to scroll.
+        XCTAssertNil(arbiter.update(centroidX: 40, centroidY: 40, distance: 100))
+        XCTAssertEqual(arbiter.mode, .zoom)
+    }
+
+    func testTwoFingerArbiterPinchWinsWhenBothCross() {
+        var arbiter = TwoFingerArbiter(centroidX: 0, centroidY: 0, distance: 100)
+        XCTAssertEqual(arbiter.update(centroidX: 15, centroidY: 0, distance: 130), .zoom)
+        XCTAssertEqual(arbiter.mode, .zoom)
+    }
+
+    func testTwoFingerArbiterIgnoresJitter() {
+        var arbiter = TwoFingerArbiter(centroidX: 0, centroidY: 0, distance: 100)
+        XCTAssertNil(arbiter.update(centroidX: 4, centroidY: 3, distance: 95))
+        XCTAssertNil(arbiter.update(centroidX: -2, centroidY: 6, distance: 108))
+        XCTAssertEqual(arbiter.mode, .undetermined)
     }
 
     func testMomentumGeneratorDecaysToFinish() {